fix: space defs: Adjust ram and register space size
Per "4.3. Space Definitions" of "SLEIGH: A Language for Rapid
Processor Specification", the size of a space is how many bytes it
takes to address each byte of the space. I know that while the H8/538
only goes up to H'FFFF, the H8/539 goes up to H'1FFFF, so mark the
size of the RAM space as 3 (could hold up to H'FFFFFF) instead of
2 (H'FFFF).

For registers: there are only a handful of registers -- their content
can be mapped entirely within 256 bytes of address space, so that
space only needs size=1. I choose not to wordsize=2 that space
here (as I could per 4.3), because I can't do pointer arithmetic on
registers anyways -- that is explicitly disallowed for a
register_space.

Also, get rid of the definition for a space called 'address'. This was
meant to be the whole address space -- RAM and ROM -- but the
rationale for separating them (so that the rest of the specification
would be forced to refer to one space or the other) was incorrect.